requirements.txt
# Core dependencies
playwright>=1.40.0
gspread>=5.12.0  # Google Sheets API with email/password support
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0

# Database
# SQLite is built-in to Python

# Utilities
python-dateutil>=2.8.2
pyyaml>=6.0.1
keyring>=24.3.0  # For secure credential storage

# Development (optional)
pytest>=7.4.0
black>=23.11.0

